Moves updated "Record" app to pre-existing slug.
Why
 ---
The slug wasn't aligned with the published version: causing two versions of the same application when published to Expo.
By adopting the old `/record` slug instead of `/audio-recording-example`, there are fewer active slugs to manage, and it allows those who actively use it to get the newest update without migrating to a new instance of the record app.

How
 ---
I updated the `"slug"` property in `app.json` to reflect the proper location of this app.

Test Plan
 ---
I visited https://expo.io/@community/record and tested the app on my Expo SDK27-enabled device to make sure the update was published successfully.

fbshipit-source-id: 426e942