Upgrade to Expo SDK v34.0.0
Why
 ---
The demo should be available to run on the latest version of Expo.

How
 ---
* I updated the sdkVersion in app.json.
* I updated the dependencies in package.json.
* I added a `scripts` section to package.json based on those added to a
  new project created with `expo init`.
* I updated App.js to fix breaking changes in the SDKs between v27 and
  v34, as per the release notes.

Test Plan
 ---
I ran the project with `yarn start` on an Android phone using Expo
client version 2.12.0.
I was prompted to allow recording, after which I tested recording and
playback functionality to ensure the correct behaviour.